Bugs fixed

+++ sensum/segmentation_opt.py
@@ -68,6 +68,9 @@ def create_extended_patch(input_band_list,reference_band,input_band_geo_transfor
     Last modified: 25/03/2014
     '''
     
+    #TODO: Please avoid os.makedirs and similar file system commands whenever possible! Again, we should work with arrays and not with files.
+    #TODO: Not clear to me what this function does from looking at the description. 
+    #TODO: What do you mean with extended raster?, Does it mask the raster to the reference objects?
     output_band_list = []
     
     rows,cols = input_band_list[0].shape
@@ -134,7 +137,8 @@ def evaluation(input_band,reference_band,input_band_geo_transform,ref_geo_transf
     Last modified: 25/03/2014
     '''
     
-    #TODO: Please add references and update function description
+    #TODO: Each function should take as input a matrix. If a file is used as input, we should use before the according conversion function. This should be done outside the function in the main file.
+    #TODO: Please add references!
     #TODO: define options for select_criteria argument (you have them in comments within the code, but we need to have clear in the description)
     #TODO: Taking the sum of all criteria is correct?
     #TODO: We should probably add another function or extend this one in order to allow for not only a supervised evaluation with reference object (as is done here), but also to allow for an unsupervised one (e.g. for Landsat data)
@@ -264,7 +268,9 @@ def bound_generator(segmentation_name):
     Last modified: 25/03/2014
     '''
     
+    #TODO: I would need a walk through the whole segmentation optimization procedure.
     #TODO: OTB segmentations are not supported by the optimization procedure?
+    #TODO: Again, would exclude Baatz and Region Growing algorithms.
     
     mybounds =[]
     
@@ -322,12 +328,12 @@ def bound_generator(segmentation_name):
         ep = 0.005
         
     if segmentation_name == 'Mprofiles':
-        min_size_bound = [0,0.95]
+        min_size_bound = [1,20]
         mybounds = [min_size_bound]
         
-        min_size = random.uniform(level_bound[0],level_bound[1])
+        min_size = random.randrange(min_size_bound[0],min_size_bound[1],2)
         parameters = [min_size]
-        ep = 0.005
+        ep = 1
         
     if segmentation_name == 'Baatz':
         compactness_bound = [0.05,0.94]
@@ -395,6 +401,7 @@ def optimizer(parameters,segmentation_name,patches_list,reference_band_list,patc
     
     
     if segmentation_name == 'Felzenszwalb':
+        print 'working...'
         for i in range(0,len(patches_list)):
             seg = felzenszwalb_skimage(patches_list[i], parameters[0], parameters[1], 0)
             eval_criteria = evaluation(seg,reference_band_list[i],patches_geo_transform_list[i],ref_geo_transform_list[i],select_criteria)
@@ -418,7 +425,7 @@ def optimizer(parameters,segmentation_name,patches_list,reference_band_list,patc
                     seg_list = read_image(path + separator + 'temp_seg.tif',np.uint16,0)
                     seg = seg_list[0]
                 if segmentation_name == 'Mprofiles':
-                    mprofiles_otb(path + separator + 'temp.tif','raster',path + separator + 'temp_seg.tif',float(parameters[0]),0,0,0)
+                    mprofiles_otb(path + separator + 'temp.tif','raster',path + separator + 'temp_seg.tif',int(parameters[0]),0,0,0)
                     seg_list = read_image(path + separator + 'temp_seg.tif',np.uint16,0)
                     seg = seg_list[0]
                 if segmentation_name == 'Baatz':
@@ -472,7 +479,9 @@ def call_optimizer(segmentation_name,patches_list,reference_band_list,patches_ge
     print 'opt_parameters',opt_parameters
     
     path = os.getcwd()
-    os.remove(path+separator+'temp.tif')
-    os.remove(path+separator+'temp_seg.tif')
+    if os.path.isfile(path+separator+'temp.tif'):
+        os.remove(path+separator+'temp.tif')
+    if os.path.isfile(path+separator+'temp_seg.tif'):
+        os.remove(path+separator+'temp_seg.tif')
     
     return opt_parameters
